Fixed Centering of Autogenerated Profile Picture
Just slapped a TextView on top of the ImageView that we already have. Now we don't have a centering problem.
Keeping TextDrawable in the back burner just in case, but probably will be deprecated in the future.
TODO: find a way to pull the user's Profile Picture from the Database