Added carry-over of eventID from OrganizerEventDetailsActivity to EventAttendeesInfoActivity
Made it so that the activity before EventAttendeesInfoActivity sends the proper eventID key in the intent.putExtra(), this way we can pull the eventID that was selected in that activity and properly display the correct users and geolocations. As well as an extra check for the conditions to ensure that we are getting user's that agreed to have their location tracked AND that they have a non-null GeoPoint object.