Java docs (PLEASE READ DESC)
These javadocs were generated without modifying any of the code in the github. I had thought that when generating them from my computer that i would automatically add into the github. If this is not permitted than you can reject this file, but all the java docs are inside the code.